Created guest-lvm-backup.rkt.
This program is fully able to, shutdown a guest operating system, snapshot its lvm, backsup the lvm image to a remote server.

It is able to fully copy the backup if one does not exist, or if need be, differentially update based upon the latest existing backup on the server.