Adds progress reporting for entire basis upload.
Adds code to remove temporary basis file on complete upload failure.

Changes location of temporary basis file to be the location that backups are stored, instead of temporary directory.

Adds changes to allow ssh to use an identity file and a port argument.

Adds progress reporting when performing a full backup.

Adds copy-port-progress method which outputs progress to a designated port when copying ports.